Moved theme logic into component. Added localstorage
